refactor integration tests into own dir
