+++ gpMgmt/go-utils/src/gp_upgrade/compile_test_sshd_server.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-pushd "./commands/sshd" 1>/dev/null
+pushd "./integrations/sshd" 1>/dev/null
   go get
   go build -o "$GOPATH/bin/test/sshd"
   RET=$?
