+++ gpMgmt/go-utils/src/gp_upgrade/integrations/all_test.go
@@ -1,4 +1,4 @@
-package commands_test
+package integrations_test
 
 import (
 	. "github.com/onsi/ginkgo"
