+++ gpMgmt/go-utils/src/gp_upgrade/integrations/fixtures/object_count.sql
