+++ gpMgmt/go-utils/src/gp_upgrade/integrations/fixtures/registered_private_key.pem
