+++ gpMgmt/go-utils/src/gp_upgrade/integrations/fixtures/registered_public_key.pem
