+++ gpMgmt/go-utils/src/gp_upgrade/integrations/fixtures/segment_config.sql
