+++ gpMgmt/go-utils/src/gp_upgrade/integrations/fixtures/unregistered_private_key.pem
