+++ gpMgmt/go-utils/src/gp_upgrade/integrations/sshd/public_key.pem
